Add files via upload
This application of predicting the heating load of the properties listed in the dataset. It is based on a regression problem, so we used forms of regression algorithms to illustrate this ML analysis process. 

The algorithms used:

Simple Linear Regression 
Ridge Regression 
Lasso Regression 
Elastic Net Regression 
Polynomial Regression with regularisation 
Neural Network

Potential Improvements:

Further feature selection and combining attributes, deleting the least collaborative features, adding parameters to the grid search cross-validation, and fine-tuning to decrease the MSE and RMSE.